Flask Application/requirements.txt
Flask==2.1.3
numpy==1.21.6
pillow==9.1.1
requests==2.28.1
google-cloud-storage